Feature nmpt 1777

+++ NMP-Portal/NMP-Portal/ServiceResponses/NitrogenApplicationsForNMaxReportResponse.cs
@@ -8,11 +8,11 @@ namespace NMP.Portal.ServiceResponses
         public string FieldName { get; set; } = string.Empty;
         public string CropTypeName { get; set; } = string.Empty;
         public decimal CropArea { get; set; }
-        public decimal? InorganicNRate { get; set; }
+        public int? InorganicNRate { get; set; }
         public int? InorganicNTotal { get; set; }
-        public decimal? OrganicCropAvailableNRate { get; set; }
+        public int? OrganicCropAvailableNRate { get; set; }
         public int? OrganicCropAvailableNTotal { get; set; }
-        public decimal? NRate { get; set; }
-        public decimal? NTotal { get; set; }
+        public int? NRate { get; set; }
+        public int? NTotal { get; set; }
     }
 }
